From Vincent Gouvernelle:   * Use uintptr_t instead of long.   * Add ASTNodeEraser to the cmake build system.
git-svn-id: http://root.cern.ch/svn/root/trunk@43480 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cling/lib/Interpreter/Interpreter.cpp
@@ -221,7 +221,7 @@ namespace cling {
       // Set up the gCling variable
       processLine("#include \"cling/Interpreter/ValuePrinter.h\"\n");
       std::stringstream initializer;
-      initializer << "gCling=(cling::Interpreter*)" << (long)this << ";";
+      initializer << "gCling=(cling::Interpreter*)" << (uintptr_t)this << ";";
       processLine(initializer.str());
     }
     else {
